reworked rounding in pubglast to show no decimal places

+++ responses/pubglast.py
@@ -49,7 +49,7 @@ class ResponsePUBGLast(AbstractResponse):
                     stats = player["attributes"]["stats"]
 
                     out = template.format(name = stats["name"],
-                                          damage = round(stats["damageDealt"], 0),
+                                          damage = int(round(stats["damageDealt"], 0)),
                                           numKills = stats["kills"],
                                           killRank = stats["killPlace"],
                                           result = stats["winPlace"],
